+++ src/main/java/com/geocode/search/cli/Parameters.java
@@ -1,82 +1,84 @@
 package com.geocode.search.cli;
 
-import static com.geocode.search.cli.Descriptions.*;
+import static com.geocode.search.message.Alert.*;
 
-import com.geocode.search.cli.settings.FileConfiguration;
-import com.geocode.search.cli.settings.IntersectModel;
-import com.geocode.search.cli.settings.IntersectParameters;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.geocode.search.logging.Logger;
+import com.geocode.search.settings.FileSettings;
+import com.geocode.search.settings.IntersectParams;
+import com.geocode.search.settings.IntersectSettings;
+import com.geocode.search.yaml.YamlStructure;
+import io.sentry.Sentry;
 import java.io.*;
-import java.util.Properties;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.apache.commons.cli.CommandLine;
 
-public class Parameters {
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Parameters extends Logger {
 
-	private FileConfiguration fileConfiguration = new FileConfiguration();
-	private IntersectModel intersectModel = new IntersectModel();
-	private IntersectParameters intersectParameters = new IntersectParameters();
+	private FileSettings fileSettings = new FileSettings();
+	private IntersectSettings intersectSettings = new IntersectSettings();
+	private IntersectParams intersectParams = new IntersectParams();
 	private int threads = 1;
-
-	public FileConfiguration getFileConfiguration() {
-		return fileConfiguration;
-	}
-
-	public void setFileConfiguration(FileConfiguration fileConfiguration) {
-		this.fileConfiguration = fileConfiguration;
-	}
-
-	public IntersectModel getIntersectModel() {
-		return intersectModel;
-	}
-
-	public void setIntersectModel(IntersectModel intersectModel) {
-		this.intersectModel = intersectModel;
-	}
-
-	public IntersectParameters getIntersectParameters() {
-		return intersectParameters;
-	}
-
-	public void setIntersectParameters(IntersectParameters intersectParameters) {
-		this.intersectParameters = intersectParameters;
-	}
-
-	public int getThreads() {
-		return threads;
-	}
-
-	public void setThreads(int threads) {
-		this.threads = threads;
-	}
 
 	/**
 	 * Method used to read parameters from the configuration file
 	 * @param cmd command line
-	 * @return boolean value that certifies whether the parameter reading was successful
+	 * @return boolean value indicating whether the parameter is correct
 	 */
-	public boolean readParams(CommandLine cmd) {
-		try {
-			if (cmd.getOptionValue("thread") != null) {
-				threads = Integer.parseInt(cmd.getOptionValue("thread"));
-			}
-			String configFilePath = cmd.getOptionValue("config");
-			FileInputStream fileConfig = new FileInputStream(configFilePath);
-			Properties properties = new Properties();
-			properties.load(fileConfig);
+	public boolean readInputParameters(CommandLine cmd) {
+		if (cmd.getOptionValue("thread") != null) {
+			threads = Integer.parseInt(cmd.getOptionValue("thread"));
+		}
 
-			if (fileConfiguration.readConfigFromProperties(properties)) {
-				if (intersectModel.readConfigFromProperties(properties)) {
-					return intersectParameters.readSettingsFromProperties(properties);
-				}
-			}
-			return false;
+		String configFilePath = cmd.getOptionValue("config");
+		YamlStructure yamlStructure = readYaml(configFilePath);
 
-		} catch (FileNotFoundException e) {
-			System.out.println(FILE_CONFIG_INVALID.description);
-			return false;
-		} catch (IOException e) {
-			System.out.println(ERROR_READ_CONFIG.description);
+		if (yamlStructure != null) {
+			return readConfigFromYaml(yamlStructure);
+		} else {
 			return false;
 		}
+	}
+
+	/**
+	 * Method used to read the .yaml configuration file
+	 * @param path configuration file path
+	 * @return data contained in the configuration file
+	 */
+	private YamlStructure readYaml(String path) {
+		try {
+			ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+			mapper.findAndRegisterModules();
+			return mapper.readValue(new File(path), YamlStructure.class);
+		} catch (Exception e) {
+			printError(ERROR_READ_CONFIG.description, e.getMessage());
+			Sentry.captureException(e);
+			return null;
+		}
+	}
+
+	/**
+	 * Method used to read yaml configuration parameters
+	 * @param yaml configuration properties
+	 * @return boolean value indicating whether the parameter is correct
+	 */
+	private boolean readConfigFromYaml(YamlStructure yaml) {
+		boolean correctSettings = fileSettings.readConfigFromYaml(yaml);
+		if (correctSettings) {
+			correctSettings = intersectSettings.readConfigFromYaml(yaml);
+		}
+		if (correctSettings) {
+			intersectParams.readConfigFromYaml(yaml);
+		}
+		return correctSettings;
 	}
 
 	/**
@@ -84,13 +86,14 @@ public class Parameters {
 	 */
 	public void closeAllConnection() {
 		try {
-			fileConfiguration.getInputFile().close();
-			fileConfiguration.getOutputFile().close();
-			if (intersectModel.getDatabaseConnection() != null) {
-				intersectModel.getDatabaseConnection().closeConnection();
+			fileSettings.getInputFile().close();
+			fileSettings.getOutputFile().close();
+			if (intersectSettings.getDatabaseConnection() != null) {
+				intersectSettings.getDatabaseConnection().closeConnection();
 			}
 		} catch (Exception e) {
-			System.out.println("Error while closing all connections. Description: " + e.getMessage());
+			printError(ERROR_CLOSING_CONNECTIONS.description, e.getMessage());
+			Sentry.captureException(e);
 		}
 	}
 }