+++ src/main/java/com/geocode/search/service/intersect/GeoTool.java
@@ -1,15 +1,22 @@
 package com.geocode.search.service.intersect;
 
-import com.geocode.search.cli.settings.IntersectParameters;
+import static com.geocode.search.message.Alert.*;
+
 import com.geocode.search.connection.Database;
+import com.geocode.search.logging.Logger;
 import com.geocode.search.service.input.Candidate;
 import com.geocode.search.service.intersect.shapefile.ShapeData;
 import com.geocode.search.service.output.IntersectResult;
+import com.geocode.search.settings.IntersectParams;
+import io.sentry.Sentry;
 import java.io.File;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Vector;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.geotools.data.DataUtilities;
 import org.geotools.data.FileDataStore;
 import org.geotools.data.FileDataStoreFinder;
@@ -21,50 +28,27 @@ import org.geotools.filter.text.cql2.CQL;
 import org.opengis.feature.simple.SimpleFeature;
 import org.opengis.feature.simple.SimpleFeatureType;
 
-public class GeoTool {
+@Getter
+@Setter
+@NoArgsConstructor
+public class GeoTool extends Logger {
 
 	private static SimpleFeatureSource source;
 	private static SimpleFeatureType schema;
 	private static final Vector<SimpleFeatureSource> cachedSourceVect = new Vector<>();
 	private static final Vector<SimpleFeatureType> schemaVect = new Vector<>();
 
-	public static SimpleFeatureSource getSource() {
-		return source;
-	}
-
-	public static void setSource(SimpleFeatureSource source) {
-		GeoTool.source = source;
-	}
-
-	public static SimpleFeatureType getSchema() {
-		return schema;
-	}
-
-	public static void setSchema(SimpleFeatureType schema) {
-		GeoTool.schema = schema;
-	}
-
-	public static Vector<SimpleFeatureSource> getCachedSourceVect() {
-		return cachedSourceVect;
-	}
-
-	public static Vector<SimpleFeatureType> getSchemaVect() {
-		return schemaVect;
-	}
-
-	public GeoTool() {}
-
 	/**
 	 * Constructor
 	 * @param fileName shapefile name
 	 * @param activeCache cache enable flag
 	 */
+	// spotless:off
 	public GeoTool(String fileName, String activeCache) {
 
 		File file = new File(fileName);
 		try {
-			System.out.println("- Start upload shapefile " + file.getName() + ": "
-					+ Calendar.getInstance().getTime());
+			printInfo("- Start upload shapefile " + file.getName() + ": " + Calendar.getInstance().getTime());
 			FileDataStore myData = FileDataStoreFinder.getDataStore(file);
 			source = myData.getFeatureSource();
 			schema = source.getSchema();
@@ -76,15 +60,16 @@ public class GeoTool {
 				cachedSourceVect.add(source);
 				schemaVect.add(schema);
 			}
-			System.out.println("- End upload shapefile " + file.getName() + ": "
-					+ Calendar.getInstance().getTime());
+			printInfo("- End upload shapefile " + file.getName() + ": " + Calendar.getInstance().getTime());
 
 		} catch (Exception e) {
-			System.out.println("- Error during upload shapefile: " + e.getMessage());
+			printError(ERROR_UPLOAD_SHAPEFILE.description, e.getMessage());
+			Sentry.captureException(e);
 			cachedSourceVect.add(source);
 			schemaVect.add(schema);
 		}
 	}
+	// spotless:on
 
 	/**
 	 * Method used to project coordinates onto the shapefile
@@ -92,15 +77,12 @@ public class GeoTool {
 	 * @param intersectResult area to be enhanced
 	 * @return candidate extracted from the shapefile
 	 */
+	// spotless:off
 	public IntersectResult intersectShapefile(Candidate candidate, IntersectResult intersectResult) {
 
 		ShapeData shapeData = new ShapeData();
 		try {
-			shapeData.setFilter(CQL.toFilter("intersects(the_geom, POINT("
-					+ candidate.getCoordinateX()
-					+ " "
-					+ candidate.getCoordinateY()
-					+ "))"));
+			shapeData.setFilter(CQL.toFilter("intersects(the_geom, POINT(" + candidate.getCoordinateX() + " " + candidate.getCoordinateY() + "))"));
 
 			for (int i = 0; i < cachedSourceVect.size(); i++) {
 				synchronized (cachedSourceVect) {
@@ -110,76 +92,77 @@ public class GeoTool {
 					shapeData.setCollection(cachedSourceVect.get(i).getFeatures(shapeData.getQuery()));
 				}
 			}
-			try (FeatureIterator<SimpleFeature> features =
-					shapeData.getCollection().features()) {
+
+			try (FeatureIterator<SimpleFeature> features = shapeData.getCollection().features()) {
 				while (features.hasNext()) {
 					intersectResult.getShapeElements().add(features.next());
 				}
 			}
 
 		} catch (Exception e) {
-			System.out.println("Error during intersect shapefile: " + e.getMessage());
+			printError(ERROR_INTERSECT_SHAPEFILE.description, e.getMessage());
+			Sentry.captureException(e);
 			System.exit(1);
 		}
 		return intersectResult;
 	}
+	// spotless:on
 
 	/**
 	 * Method used to project coordinates onto the shapefile and extract n candidates
 	 * @param candidate candidate to be projected on the shapefile
-	 * @param intersectParameters settings used for research
+	 * @param intersectParams settings used for research
 	 * @return candidates extracted from the shapefile
 	 */
-	public IntersectResult extractDataFromShapefile(Candidate candidate, IntersectParameters intersectParameters) {
+	// spotless:off
+	public IntersectResult extractDataFromShapefile(Candidate candidate, IntersectParams intersectParams) {
 
 		IntersectResult intersectResult = new IntersectResult();
 		try {
 			double lap = 1;
 			double distance = 0;
-			double increase = 0.00001 * intersectParameters.getIncrease();
+			double increase = 0.00001 * intersectParams.getIncrease();
 			intersectResult = intersectShapefile(candidate, intersectResult);
 
-			for (int i = 1; i < intersectParameters.getAttempts(); i++) {
-				if (intersectResult.getShapeElements().size() == intersectParameters.getCandidates()) break;
-				if (distance >= intersectParameters.getMaxDistance()) break;
+			for (int i = 1; i < intersectParams.getAttempts(); i++) {
+				if (intersectResult.getShapeElements().size() == intersectParams.getCandidates()) break;
+				if (distance >= intersectParams.getMaxDistance()) break;
 
-				for (double x = candidate.getCoordinateX() - increase;
-						x <= candidate.getCoordinateX() + increase;
-						x += increase / lap) {
+				for (double x = candidate.getCoordinateX() - increase;	x <= candidate.getCoordinateX() + increase;	x += increase / lap) {
 
-					for (double y = candidate.getCoordinateY() - increase;
-							y <= candidate.getCoordinateY() + increase;
-							y += increase / lap) {
+					for (double y = candidate.getCoordinateY() - increase; y <= candidate.getCoordinateY() + increase; y += increase / lap) {
 
 						Candidate point = new Candidate(x, y);
 						intersectResult = intersectShapefile(point, intersectResult);
-						if (intersectResult.getShapeElements().size() == intersectParameters.getCandidates()) break;
+						if (intersectResult.getShapeElements().size() == intersectParams.getCandidates()) break;
 					}
 
-					if (intersectResult.getShapeElements().size() == intersectParameters.getCandidates()) break;
+					if (intersectResult.getShapeElements().size() == intersectParams.getCandidates()) break;
 				}
 
-				distance += intersectParameters.getIncrease();
-				increase += 0.00001 * intersectParameters.getIncrease();
+				distance += intersectParams.getIncrease();
+				increase += 0.00001 * intersectParams.getIncrease();
 				lap++;
 			}
 		} catch (Exception e) {
-			System.out.println("Error when extracting data from shapefile: " + e.getMessage());
+			printError(ERROR_EXTRACT_DATA_SHAPEFILE.description, e.getMessage());
+			Sentry.captureException(e);
 			System.exit(1);
 		}
 		return intersectResult;
 	}
+	// spotless:on
 
 	/**
 	 * Method used to project the coordinates onto the database and extract the specified columns
 	 * @param candidate candidate to be projected on the database
 	 * @param database database connection
+	 * @param limit number of candidates to extract
 	 * @param columns database columns
 	 * @return list of candidates close to the given point
 	 */
-	public IntersectResult extractDataFromDatabase(
-			Candidate candidate, Database database, double limit, ArrayList<String> columns) {
-
+	// spotless:off
+	public IntersectResult extractDataFromDatabase(Candidate candidate, Database database, double limit, ArrayList<String> columns) {
 		String query = "SELECT DISTINCT ";
 		for (String column : columns) {
 			query += column + ",";
@@ -200,12 +183,13 @@ public class GeoTool {
 
 		return executeIntersect(database, query);
 	}
+	// spotless:on
 
 	/**
 	 * Method used to project coordinates onto the database and extract n candidates given in the limit parameter
 	 * @param candidate candidate to be projected on the database
 	 * @param database database connection
-	 * @param limit maximum number of candidates to be drawn
+	 * @param limit number of candidates to extract
 	 * @return list of candidates close to the given point
 	 */
 	public IntersectResult extractDataFromDatabase(Candidate candidate, Database database, double limit) {
@@ -241,11 +225,10 @@ public class GeoTool {
 			if (database.getConnection() != null) {
 				Statement stmt = database.getConnection().createStatement();
 				intersectResult.setDbElements(stmt.executeQuery(query));
-			} else {
-				System.out.println("Error during database connection.");
 			}
 		} catch (Exception e) {
-			System.out.println("Error during intersect database: " + e.getMessage());
+			printError(ERROR_INTERSECT_DATABASE.description, e.getMessage());
+			Sentry.captureException(e);
 		}
 		return intersectResult;
 	}