+++ src/main/java/com/geocode/search/service/output/IntersectResult.java
@@ -2,26 +2,18 @@ package com.geocode.search.service.output;
 
 import java.sql.ResultSet;
 import java.util.HashSet;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.opengis.feature.simple.SimpleFeature;
 
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
 public class IntersectResult {
 
 	private HashSet<SimpleFeature> shapeElements = new HashSet<>();
 	private ResultSet dbElements;
-
-	public HashSet<SimpleFeature> getShapeElements() {
-		return shapeElements;
-	}
-
-	public void setShapeElements(HashSet<SimpleFeature> shapeElements) {
-		this.shapeElements = shapeElements;
-	}
-
-	public ResultSet getDbElements() {
-		return dbElements;
-	}
-
-	public void setDbElements(ResultSet dbElements) {
-		this.dbElements = dbElements;
-	}
 }