+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## Version 1.0.1
+### 2024-10-07
+- Added exception catching via Sentry
+- Sentry 4.11.0
+
+### 2024-10-04
+- Log message handling via Log4j2
+- Log4j2 2.24.1
+
+### 2024-09-29
+- Lombok 1.18.34
+- Jackson-dataformat-yaml 2.18.0
+- Added .yaml as configuration file
+
 ## Version 1.0.0
 ### 2024-06-26
 - Initial release